Move SerializablePropertyAttribute to its own file
